+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - 
 ### Fixed:
 - Fixed blocking animation
+- Fixed entity crash
 ### Changed:
 - Removed dependency on ARRP/BRRP
 - Applies when non compat mods are loaded